+++ app/src/main/java/com/marelso/designsystem/navigation/NavGraph.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.marelso.designsystem.ui.home.components.HomeScreenHoisting
+import com.marelso.designsystem.ui.home.HomeScreenHoisting
 
 @Composable
 fun NavGraph(navHostController: NavHostController, onClose: () -> Unit) {
