+++ app/src/main/java/com/marelso/designsystem/ui/home/HomeScreenHoisting.kt
@@ -1,4 +1,4 @@
-package com.marelso.designsystem.ui.home.components
+package com.marelso.designsystem.ui.home
 
 import androidx.compose.runtime.Composable
